Added JWT token validation to only allow Sellers and Admins to create, update, delete or replace a product

+++ src/main/java/com/vibevault/productservice/controllers/ProductController.java
@@ -1,11 +1,15 @@
 package com.vibevault.productservice.controllers;
 
+import com.vibevault.productservice.commons.AuthenticationCommons;
+import com.vibevault.productservice.dtos.commons.UserDto;
+import com.vibevault.productservice.dtos.exceptions.authentication.InvalidTokenException;
 import com.vibevault.productservice.dtos.product.*;
 import com.vibevault.productservice.exceptions.products.ProductNotCreatedException;
 import com.vibevault.productservice.exceptions.products.ProductNotDeletedException;
 import com.vibevault.productservice.exceptions.products.ProductNotFoundException;
 import com.vibevault.productservice.models.Product;
 import com.vibevault.productservice.services.ProductService;
+import jakarta.annotation.Nullable;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,17 +20,31 @@ import java.util.UUID;
 public class ProductController {
 
     private ProductService productService;
+    private AuthenticationCommons authenticationCommons;
 
-    public ProductController(ProductService productService) {
+    public ProductController(ProductService productService,
+                             AuthenticationCommons authenticationCommons) {
         this.productService = productService;
+        this.authenticationCommons = authenticationCommons;
     }
     @PostMapping("")
-    public CreateProductResponseDto createProduct(@RequestBody CreateProductRequestDto createProductRequestDto) throws ProductNotCreatedException {
+    public CreateProductResponseDto createProduct(@RequestBody CreateProductRequestDto createProductRequestDto,
+                                                  @Nullable @RequestHeader("Authorization") String token) throws ProductNotCreatedException, InvalidTokenException {
+        UserDto userDto = authenticationCommons.validateToken(token);
+        if (userDto == null || (!userDto.getRoles().contains("SELLER") && !userDto.getRoles().contains("ADMIN"))) {
+            throw new InvalidTokenException("Unauthorized to create product");
+        }
         Product product = productService.createProduct(createProductRequestDto.toProduct());
         return CreateProductResponseDto.fromProduct(product);
     }
     @PatchMapping("/{productId}")
-    public UpdateProductResponseDto updateProduct(@PathVariable("productId") String productId, @RequestBody UpdateProductRequestDto updateProductRequestDto) throws ProductNotFoundException {
+    public UpdateProductResponseDto updateProduct(@PathVariable("productId") String productId,
+                                                  @RequestBody UpdateProductRequestDto updateProductRequestDto,
+                                                  @Nullable @RequestHeader("Authorization") String token) throws ProductNotFoundException, InvalidTokenException {
+        UserDto userDto = authenticationCommons.validateToken(token);
+        if (userDto == null || (!userDto.getRoles().contains("SELLER") && !userDto.getRoles().contains("ADMIN"))) {
+            throw new InvalidTokenException("Unauthorized to update product");
+        }
         Product product = productService.updateProduct(productId, updateProductRequestDto.toProduct());
         return UpdateProductResponseDto.fromProduct(product);
     }
@@ -41,13 +59,24 @@ public class ProductController {
         return GetProductResponseDto.fromProducts(products);
     }
     @DeleteMapping("/{productId}")
-    public DeleteProductResponseDto deleteProduct(@PathVariable("productId") String productId) throws ProductNotFoundException, ProductNotDeletedException {
+    public DeleteProductResponseDto deleteProduct(@PathVariable("productId") String productId,
+                                                  @Nullable @RequestHeader("Authorization") String token) throws ProductNotFoundException, ProductNotDeletedException {
+        UserDto userDto = authenticationCommons.validateToken(token);
+        if (userDto == null || (!userDto.getRoles().contains("SELLER") && !userDto.getRoles().contains("ADMIN"))) {
+            throw new InvalidTokenException("Unauthorized to delete product");
+        }
         Product product=productService.deleteProduct(productId);
         return DeleteProductResponseDto.fromProduct(product);
     }
 
     @PutMapping("/{productId}")
-    public ReplaceProductResponseDto replaceProduct(@PathVariable("productId") String productId, @RequestBody ReplaceProductRequestDto replaceProductRequestDto) throws ProductNotFoundException {
+    public ReplaceProductResponseDto replaceProduct(@PathVariable("productId") String productId,
+                                                    @RequestBody ReplaceProductRequestDto replaceProductRequestDto,
+                                                    @Nullable @RequestHeader("Authorization") String token) throws ProductNotFoundException {
+        UserDto userDto = authenticationCommons.validateToken(token);
+        if (userDto == null || (!userDto.getRoles().contains("SELLER") && !userDto.getRoles().contains("ADMIN"))) {
+            throw new InvalidTokenException("Unauthorized to replace product");
+        }
         Product product = productService.replaceProduct(productId, replaceProductRequestDto.toProduct());
         return ReplaceProductResponseDto.fromProduct(product);
